Improved weapon systems, fixed player input interactions with UI
Added a method to weapon:magic that sets the projectile initial offset, to fix a bug, when destructible target is affected, if it stays behind indestructible object and player shots end to end.
Added an overloaded method to update energy bars in some specific cases.
Added some checks to ensure player interaction with UI elems (rather than standart input interactions: weapon hitting, aiming etc), when cursor is on it